feat(): Organizado arquivos da pasta services, Criado novos arquivos
