+++ .env
@@ -1,1 +1,1 @@
-SERVER_URL=
+SERVER_URL=https://softeo-tecnologia-backend.herokuapp.com
